refactor(validator): validator 클래스를 validator 패키지로 이동
