+++ src/main/java/lotto/domain/validator/LottoCountValidator.java
@@ -1,9 +1,9 @@
-package lotto.view;
+package lotto.domain.validator;
 
 public final class LottoCountValidator extends NumberValidator {
 
     @Override
-    protected void validate(String input) {
+    public void validate(String input) {
         checkNumeric(input);
         checkPositive(input);
     }
