+++ src/main/java/lotto/domain/validator/LottoNumberValidator.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.domain.validator;
 
 import lotto.domain.LottoNumber;
 
@@ -7,7 +7,7 @@ public final class LottoNumberValidator extends NumberValidator {
     static final String OUT_BOUND_MESSAGE = "숫자는 1 ~ 45 사이의 값을 가져야 합니다!";
 
     @Override
-    protected void validate(String input) {
+    public void validate(String input) {
         super.validate(input);
         checkBound(input);
     }
