+++ src/test/java/lotto/domain/validator/LottoCountValidatorTest.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.domain.validator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
